Initial (almost useless) implementation of doc generator.
To generate bindings doc a hack (or a patch into crystal doc) need
to be done. As crystal doc reads all .cr files it end up detecting
duplicated constants definitions if there's a constant definition in
a human generated file included in the binding.

So you need to remove the generated binding/includes/ directory to be
able to run crystal doc.

For now just basic class/struct/interface docs is print with no
processing, useless but a start.