Refactor code that deals with argument transformations.
So now it must be possible to fix https://github.com/hugopl/gtk4.cr/issues/12

Also more info is print on generated code reggarding the argument handling, so
it's easier to discover what piece of code generated what code.

Signal code isn't using this code yet, but the plan is to let it also use it.