Generate bindings for module functions.
This was removed by mistake after the refactor to use ECR files. Things
like this are harder to happen now since we can run `make compare` to compare
the API changes.

Note: There's a log of functions in GLib and GObject that still need to be
ignored... I just added all now because this is boring as hell.