Add basic stuff to be able to subclass GObject.
For now just the new type is registered on GLib GObject type system.

The plan is to use a custom structure for these type with a pointer
to the Crystal object, so when the object goes to C world and back
to Crystal world we can retrieve the Crystal object.

Need to check if the evil GC will see the Crystal object pointers
in the C managed memory...

Relates to issue #5.